Test fixes. Pull out upgrade for material nodes.
